pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "homie-ai"
version = "0.2.0"
description = "Fully local, privacy-first personal AI assistant with background intelligence"
readme = "README.md"
license = "MPL-2.0"
requires-python = ">=3.11"
authors = [
    { name = "MSG", email = "muthu.g.subramanian@outlook.com" },
]
keywords = ["ai", "assistant", "local", "privacy", "llm", "gguf", "llama-cpp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "cryptography>=43.0",
    "keyring>=25.0",
    "feedparser>=6.0",
    "prompt-toolkit>=3.0",
    "requests>=2.31",
    "psutil>=5.9",
]

[project.urls]
Homepage = "https://github.com/MSG-88/Homie"
Repository = "https://github.com/MSG-88/Homie"
Issues = "https://github.com/MSG-88/Homie/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]
model = [
    "transformers>=4.40",
    "accelerate>=0.30",
    "huggingface-hub>=0.23",
]
voice = [
    "faster-whisper>=1.0",
    "openwakeword>=0.6",
    "piper-tts>=1.2",
    "sounddevice>=0.4",
    "torch>=2.0",
    "kokoro>=0.9",
    "melo-tts>=0.1",
]
context = [
    "pywin32>=306; sys_platform == 'win32'",
    "psutil>=5.9",
    "watchdog>=4.0",
]
storage = [
    "chromadb>=0.5",
    "cryptography>=42.0",
]
app = [
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "pystray>=0.19",
    "Pillow>=10.0",
    "pynput>=1.7",
    "apscheduler>=3.10",
]
neural = [
    "onnxruntime>=1.17",
    "huggingface-hub>=0.23",
    "tokenizers>=0.15",
    "numpy>=1.24",
]
email = [
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
]
social = [
    "requests>=2.31",
]
messaging = ["telethon>=1.34"]
screen-reader = ["mss>=9.0", "Pillow>=10.0", "pywin32>=306"]
service = ["windows-toasts>=1.0", "plyer>=2.1"]
all = ["homie-ai[model,voice,context,storage,app,neural,email,social,messaging,screen-reader,service]"]

[project.scripts]
homie = "homie_app.cli:main"
homie-daemon = "homie_app.daemon:run_daemon"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/homie_core", "src/homie_app"]
